Support different answer types

+++ php/forms.php
@@ -29,14 +29,36 @@ class Question
     public $userId;
     public $username;
     public $answers;
-    public function __construct($id, $formId, $value, $userId, $username, $answers = [])
-    {
+    public $values;
+    public $type;
+    public $min;
+    public $max;
+    public $step;
+
+    public function __construct(
+        $id,
+        $formId,
+        $value,
+        $userId,
+        $username,
+        $values,
+        $type,
+        $min,
+        $max,
+        $step,
+        $answers = []
+    ) {
         $this->id = $id;
         $this->formId = $formId;
         $this->value = $value;
         $this->userId = $userId;
         $this->username = $username;
         $this->answers = $answers;
+        $this->values = $values;
+        $this->type = $type;
+        $this->min = $min;
+        $this->max = $max;
+        $this->step = $step;
     }
 }
 
@@ -76,12 +98,14 @@ function getForms($userId)
     return $forms;
 }
 
-function deleteForm($formId)
+function deleteForm($formId, $deleteForm = true, $useTransaction = true)
 {
     global $db;
 
     try {
-        $db->beginTransaction();
+        if ($useTransaction) {
+            $db->beginTransaction();
+        }
 
         $db->query('
         delete from answer
@@ -95,15 +119,21 @@ function deleteForm($formId)
             ['form_id' => $formId]
         );
 
-        $db->query('
-        delete from form
-        where id = :form_id',
-            ['form_id' => $formId]
-        );
+        if ($deleteForm) {
+            $db->query('
+                delete from form
+                where id = :form_id',
+                ['form_id' => $formId]
+            );
+        }
 
-        $db->commit();
+        if ($useTransaction) {
+            $db->commit();
+        }
     } catch (Exception $e) {
-        $db->rollBack();
+        if ($useTransaction) {
+            $db->rollBack();
+        }
         throw $e;
     }
 }
@@ -116,7 +146,11 @@ function handleRequest()
                 handleGetRequest();
                 break;
             case 'POST':
-                handlePostRequest();
+                if (isset($_GET['formId'])) {
+                    handlePostRequest($_GET['formId']);
+                } else {
+                    handlePostRequest();
+                }
                 break;
             case 'DELETE':
                 handleDeleteRequest();
@@ -134,30 +168,60 @@ function handleRequest()
     }
 }
 
-function createForm($title, $userId)
+function createForm($form)
 {
     global $db;
 
     $db->query('
         insert into form (title, user_id)
         values (:title, :user_id)',
-        ['title' => $title, 'user_id' => $userId]
+        ['title' => $form->title, 'user_id' => $form->userId]
     );
 
-    return new Form($db->lastInsertId(), $title, $userId, []);
+    $form->id = $db->lastInsertId();
+    return $form;
 }
 
-function createQuestion($formId, $value)
+function createQuestion($question)
 {
     global $db;
 
-    $db->query('
-        insert into question (form_id, value)
-        values (:form_id, :value)',
-        ['form_id' => $formId, 'value' => $value]
-    );
+    $encodedValues = array_map('base64_encode', $question->values);
+    $concatenatedValues = implode(',', $encodedValues);
 
-    return new Question($db->lastInsertId(), $formId, $value, 0, '', []);
+    if ($question->id != 0) {
+        $db->query('
+            update question
+            set `value` = :value, `min` = :min, `max` = :max, `step` = :step, `values` = :values, `type` = :type
+            where id = :id',
+            [
+                'value' => $question->value,
+                'min' => $question->min,
+                'max' => $question->max,
+                'step' => $question->step,
+                'values' => $concatenatedValues,
+                'type' => $question->type,
+                'id' => $question->id
+            ]
+        );
+    } else {
+        $db->query('
+        insert into question (`form_id`, `value`, `min`, `max`, `step`, `values`, `type`)
+        values (:form_id, :value, :min, :max, :step, :values, :type)',
+            [
+                'form_id' => $question->formId,
+                'value' => $question->value,
+                'min' => $question->min,
+                'max' => $question->max,
+                'step' => $question->step,
+                'values' => $concatenatedValues,
+                'type' => $question->type
+            ]
+        );
+    }
+
+    $question->id = $db->lastInsertId();
+    return $question;
 }
 
 function getForm($formId)
@@ -167,7 +231,10 @@ function getForm($formId)
     $query = $db->query('
         select f.id as form_id, f.title as form_title,
             f.user_id as form_user_id, q.id as questionId,
-            q.form_id as question_form_id, q.value as question_value
+            q.form_id as question_form_id, q.value as question_value,
+            q.type as question_type, q.min as question_min,
+            q.max as question_max, q.step as question_step,
+            q.values as question_values
         from form as f
         left join question as q on f.id = q.form_id
         where f.id = :form_id',
@@ -181,14 +248,26 @@ function getForm($formId)
 
     $questions = [];
     foreach ($results as $row) {
+        $parts = explode(',', $row["question_values"]);
+        $values = [];
+
+        foreach ($parts as $part) {
+            array_push($values, base64_decode($part));
+        }
+
         array_push(
             $questions,
             new Question(
                 $row['questionId'],
                 $row['question_form_id'],
                 $row['question_value'],
+                $row['form_user_id'],
                 '',
-                '',
+                $values,
+                $row['question_type'],
+                $row['question_min'],
+                $row['question_max'],
+                $row['question_step'],
                 []
             )
         );
@@ -203,7 +282,10 @@ function getQuestion($questionId)
 
     $query = $db->query('
         select q.id as question_id, q.form_id as question_form_id,
-            q.value as question_value, u.id as user_id,
+            q.value as question_value,
+            q.type as question_type, q.min as question_min,
+            q.max as question_max, q.step as question_step,
+            q.values as question_values, u.id as user_id,
             u.username as user_username, a.id as answer_id,
             a.value as answer_value
         from question as q
@@ -236,33 +318,81 @@ function getQuestion($questionId)
         $results[0]['question_id'],
         $results[0]['question_form_id'],
         $results[0]['question_value'],
-        '',
-        '',
+        $results[0]['user_id'],
+        $results[0]['user_username'],
+        $results[0]['question_values'],
+        $results[0]['question_type'],
+        $results[0]['question_min'],
+        $results[0]['question_max'],
+        $results[0]['question_step'],
         $answers
     );
 }
 
-function handlePostRequest()
+function updateForm($form)
+{
+    global $db;
+
+    $db->query('
+        update form
+        set title = :title
+        where id = :id',
+        ['title' => $form->title, 'id' => $form->id]
+    );
+}
+
+function handlePostRequest($formId = null)
 {
     $data = json_decode(file_get_contents('php://input'), true);
-    if (
-        isset($data['userId']) && isset($data['title']) && isset($data['questions'])
-        && $data['userId'] == $_SESSION['userId']
-    ) {
-        $userId = $data['userId'];
+    if (!isset($_SESSION['userId'])) {
+        http_response_code(401);
+        echo json_encode(['error' => 'Invalid request. You are not logged in.']);
+        return;
+    } elseif (isset($data['title']) && isset($data['questions'])) {
+        $userId = $_SESSION['userId'];
         $title = $data['title'];
+        $form = new Form(0, $title, $userId, []);
 
         global $db;
         try {
             $db->beginTransaction();
 
-            $form = createForm($title, $userId);
+            if (isset($formId)) {
+                $form->id = $formId;
+                $existingForm = getForm($formId);
+
+                if (!isset($existingForm)) {
+                    $form = createForm($form);
+                } elseif ($existingForm->userId != $userId) {
+                    http_response_code(400);
+                    echo json_encode(['error' => 'Invalid request. This is not your form.']);
+                    return;
+                } else {
+                    updateForm($form);
+                }
+            } else {
+                $form = createForm($form);
+            }
 
             $questions = [];
             foreach ($data['questions'] as $question) {
                 if (isset($question['value'])) {
-                    $question = createQuestion($form->id, $question['value']);
-                    array_push($questions, $question);
+
+                    $questionObj = new Question(
+                        isset($question['id']) ? $question['id'] : 0,
+                        $form->id,
+                        $question['value'],
+                        $userId,
+                        '',
+                        $question['values'],
+                        $question['type'],
+                        $question['min'],
+                        $question['max'],
+                        $question['step'],
+                        []
+                    );
+                    $questionObj = createQuestion($questionObj);
+                    array_push($questions, $questionObj);
                 }
             }
             $form->questions = $questions;
